backend/requirements.txt
# 科学计算核心库
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# 机器学习库
scikit-learn>=1.3.0
gplearn>=0.4.2
sympy>=1.12

# 符号回归专用库
PySR>=0.12.0

# Web服务框架
flask>=2.3.0
flask-cors>=4.0.0
fastapi>=0.100.0
uvicorn>=0.23.0

# 数据处理
openpyxl>=3.1.0
xlrd>=2.0.1

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# 工具库
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0

# 开发工具
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0 